backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-socketio==5.11.1
pydantic==2.5.3
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
imagehash==4.3.1
ffmpeg-python==0.2.0
pymediainfo==6.1.0
xxhash==3.4.1
sqlalchemy==2.0.27
send2trash==1.8.3
aiofiles==23.2.1
